cmd/giouiorg: redirect Go package URLs to godoc.org
Replace the clumsy /pkg/<package> URLs with a check with godoc.org
to see whether the URL is a package in one of the gioui.org modules.
If godoc.org returns a OK (200), redirect the request.

The result is that pages that exist in the content directory are
served directly:

	https://gioui.org
	https://gioui.org/doc/install

while URLS that are valid Go import paths redirect to godoc.org:

	https://gioui.org/unit
	https://gioui.org/app/internal/window

The HEAD check ensures that URLs that are not on the content director
and also not a valid Go package will result in a 404:

	https://gioui.org/blah

Signed-off-by: Elias Naur <mail@eliasnaur.com>