cmd/giouiorg: attempt to fix meta for x/{haptic,notify}
Currently, go is failing to resolve these becaues it's always looking at
the root directory of the repository. I haven't been able to find any
documentation indicating that this subdirectory syntax is legal in
the repoRoot field, but I did see one example of it in the wild. The only
way to find out is to try it!

Signed-off-by: Chris Waldon <christopher.waldon.dev@gmail.com>