content/showcase: add example applications
This adds showcase section for the site. There are other modifications
to support this.

Add support for per page templates to create dynamic pages --
e.g. the showcase index.

Add support for serving images from the content folder. This allows to
keep the images related to the content near the content itself.

Signed-off-by: Egon Elbre <egonelbre@gmail.com>